Clean up backend documentation and header layout
