+++ libcxx/include/__pstl/backends/libdispatch.h
@@ -24,7 +24,6 @@
 #include <__memory/unique_ptr.h>
 #include <__numeric/reduce.h>
 #include <__pstl/backend_fwd.h>
-#include <__pstl/configuration_fwd.h>
 #include <__pstl/cpu_algos/any_of.h>
 #include <__pstl/cpu_algos/cpu_traits.h>
 #include <__pstl/cpu_algos/fill.h>
